requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2
django-rest-knox==4.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
pytz==2021.1
sqlparse==0.4.1
